+++ Testing.Client/Helpers/NumberFormatter.cs
@@ -4,14 +4,14 @@ namespace Testing.Client.Helpers
 {
     public class NumberFormatter: INumberFormatter
     {
-        public int GetFormattedNumber(string providedNumber)
+        public long GetFormattedNumber(string providedNumber)
         {
             // remove all non-numeric characters from the provided number value using regular expressions
-            var formattedNumber = Convert.ToInt32(Regex.Replace(providedNumber, "[^0-9]", ""));
+            var formattedNumber = Convert.ToInt64(Regex.Replace(providedNumber, "[^0-9]", ""));
 
             // ensure that the number supplied begins with a "1" prefix. If not, prepend it
             if (Convert.ToString(formattedNumber)[0] != '1')
-                formattedNumber = Convert.ToInt32('1' + Convert.ToString(formattedNumber));
+                formattedNumber = Convert.ToInt64('1' + Convert.ToString(formattedNumber));
 
             return formattedNumber;
         }
